move RDMAInterface models out of versioned API
