+++ src/middlewared/middlewared/api/v25_10_0/__init__.py
@@ -80,7 +80,6 @@ from .pool_snapshot_count import *
 from .pool_snapshottask import *
 from .privilege import *
 from .rdma import *
-from .rdma_interface import *
 from .replication import *
 from .replication_config import *
 from .replication_crud import *
